requirements.txt
Flask==0.10.1
Jinja2==2.7.3
Werkzeug==0.10.4
oauthlib==0.7.2
requests-oauthlib==0.6.0
oauth2==1.5.211
oauth2client==1.5.2
Authomatic==0.1.0.post1
gunicorn==19.4.1
flask_oauthlib
python-openid==2.2.5
python-firebase
git+https://github.com/dgrtwo/ParsePy.git
